local: Make sources and renderers deal with non-raw video
This adds a parser in front of video sources, and a decoder in video
renderers, so that if we negotiate a non-raw video format, things work
correctly.

The parser is mostly required since not all devices produce buffers that
are properly tagged with the appropriate flags (such as DELTA_UNIT).